Update setting_up_git.md; In Step 2.4 line 200
In step 2.4 linking ssh key with GitHub, to copy the SSH using "cat" command and copying SSH key, there is disperency in given instructions in line 200. For beginners there's a chance of misunderstanding and copying the whole output of "cat" command including email address. So added a note stating that, exclude email address and what happens if email address is included while adding SSH key in GitHub.